requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
async-timeout==5.0.1
attrs==24.3.0
black==24.10.0
cffi==1.17.1
click==8.1.8
colorama==0.4.6
discord.py==2.4.0
duckdb==1.1.3
frozenlist==1.5.0
idna==3.10
multidict==6.1.0
mypy-extensions==1.0.0
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
propcache==0.2.1
pycparser==2.22
PyNaCl==1.5.0
pytz==2024.2
typing_extensions==4.12.2
wavelink @ git+https://github.com/PythonistaGuild/Wavelink.git@0a39dcfa5c536c453867dc2a3ced8b3d3c6636e1
yarl==1.18.3
